+++ Changelog.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.8] - 2024-05-25
+
+### Summary
+
+- Work with framework OriCommJS 1.0.9 and above.
+
+### Added
+
+### Changed
+
+- Update example and encourage apply await response.inspector function into the gui,api,app and common/model. Completed on 2024-05-25
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+[1.0.8]: https://github.com/wkloh76/oricommjs_components/releases/tag/1.0.8
+
 ## [1.0.7] - 2024-05-16
 
 ### Summary